requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.3
MarkupSafe==1.1.1
Werkzeug==0.16.0
gunicorn
azure-identity==1.20.0
azure-keyvault-secrets==4.9.0
mysql-connector-python==8.0.23
python-dotenv==1.0.1	